Implemented saving locally to Log items

+++ app/src/main/java/rodrigues/henrique/myapplication2/InputFilterMinMax.java
@@ -4,22 +4,22 @@ import android.text.InputFilter;
 import android.text.Spanned;
 
 public class InputFilterMinMax implements InputFilter {
-    private int min, max;
+    private double min, max;
 
-    public InputFilterMinMax(int min, int max) {
+    public InputFilterMinMax(double min, double max) {
         this.min = min;
         this.max = max;
     }
 
     public InputFilterMinMax(String min, String max) {
-        this.min = Integer.parseInt(min);
-        this.max = Integer.parseInt(max);
+        this.min = Double.parseDouble(min);
+        this.max = Double.parseDouble(max);
     }
 
     @Override
     public CharSequence filter(CharSequence source, int start, int end, Spanned dest, int dStart, int dEnd) {
         try {
-            int input = Integer.parseInt(dest.toString() + source.toString());
+            double input = Double.parseDouble(dest.toString() + source.toString());
             if (isInRange(min, max, input)) {
                 return null;
             }
@@ -28,7 +28,7 @@ public class InputFilterMinMax implements InputFilter {
         return "";
     }
 
-    private boolean isInRange(int a, int b, int c) {
+    private boolean isInRange(double a, double b, double c) {
         return b > a ? c>= a && c <= b : c >= b && c<= a;
     }
 }
